Check to see if from returns a NULL pointer
If the message file is missing, any header queries will return NULL.
We need to check the first header we request and bail if it is NULL.

The file may be missing for a variety of reasons, but that is neither
here not there.  In the future, we may want to question what to do
with the database (delete the document, I think) and who is
responsible to do that.

Signed-off-by: Vernon Mauery <vernon@mauery.com>