Look harder for missing message files
Message files may be missing for a variety of reasons, the most
common (in a Maildir setup) is changing flags and/or changing
directories.  In Maildir, messages move from tmp/ to new/ to
cur/.  Depending on where it was when it was indexed, it may
be somewhere else because the mail client moved it (as per the
Maildir spec).  Also, Maildir messages have flags to specify
the status of the message, DFPRST, or Draft, Flagged, Passed
(bounced/forwarded), Replied, Seen, and Trash.  Any of these
flags may be present, but they must be in alphabetical order.

This patch looks a little harder, looking in both new/ and cur/
and testing the various combinations of flags, with the most
common situations first to speed up the process.  Then, when
it finds the file, it updates the database with the new filename.

Signed-off-by: Vernon Mauery <vernon@mauery.com>