Add thread_id as a value in the database, not just a term
Come search time, values can be forced to be unique, so we
can search for a page of threads and show those instead of
searching for a bunch of messages and then loading each thread
until we get a page.

This may cause redundant database information, but I am not
to the point that I can tell about that yet.

Signed-off-by: Vernon Mauery <vernon@mauery.com>