Initial commit of the actual packrat UI
This is really not much of a UI.  Pardon the pun.  The application
will start up a search of tag:inbox and display the results with
a cursor that moves up and down.  It will not open the message or
perform any other action on it.  It will segfault if you scroll
past the top or bottom.  But it is a start and contains the
infrastructure for building the rest of the application.  Or maybe
just a good idea of how it will go.

Signed-off-by: Vernon Mauery <vernon@mauery.com>