Add a settings class
The settings class (when complete) will expose a key/value
interface that allows for defaults if the key is not set.
It is instantiated with the argc and argv of the program
which allows it to read in the config file (from the default
path or from the command line argument) so it can read in the
config file values.  Then the command line args will be set
so they override any config file settings.

Signed-off-by: Vernon Mauery <vernon@mauery.com>