Added an editor screen (which can edit mail files)
This large patch set adds an editor screen (currently hard-coded
as vim) which allows you to edit a reply, and eventually, forwards
and new mail.  The flow is that from some screen, a message is
selected and message::reply() is called, which returns the filename
the reply is saved in.  Then editor_screen::create(filename) opens
the file for editing.  The next logical step would be a viewer screen
that can view the file and then send it.

Signed-off-by: Vernon Mauery <vernon@mauery.com>